+++ velox/functions/sparksql/tests/ParseUrlTest.cpp
@@ -24,19 +24,23 @@ using namespace facebook::velox::test;
 // All expectations below were generated by running Spark 3.5.1's
 // ParseUrl (java.net.URI based) against the same inputs, so they pin the
 // implementation to Java semantics rather than RFC 3986.
-class ParseUrlTest : public SparkFunctionBaseTest {};
-
-TEST_F(ParseUrlTest, allParts) {
-  const auto parseUrl = [&](const std::optional<std::string>& url,
-                            const std::string& part) {
+class ParseUrlTest : public SparkFunctionBaseTest {
+ public:
+  std::optional<std::string> parseUrl(
+      const std::optional<std::string>& url,
+      const std::string& part) {
     return evaluateOnce<std::string>(
         fmt::format("parse_url(c0, '{}')", part), url);
-  };
-  const auto parseUrlQuery = [&](const std::optional<std::string>& url,
-                                 const std::optional<std::string>& key) {
-    return evaluateOnce<std::string>("parse_url(c0, 'QUERY', c1)", url, key);
-  };
+  }
 
+  std::optional<std::string> parseUrlQuery(
+      const std::optional<std::string>& url,
+      const std::optional<std::string>& key) {
+    return evaluateOnce<std::string>("parse_url(c0, 'QUERY', c1)", url, key);
+  }
+};
+
+TEST_F(ParseUrlTest, allParts) {
   // A URL that carries every component Spark recognizes. Mirrors the first
   // case in Spark's UrlFunctionsSuite.
   const auto fullUrl = "http://userinfo@spark.apache.org/path?query=1#Ref";
@@ -140,12 +144,6 @@ TEST_F(ParseUrlTest, allParts) {
 // Covers empty input and scheme-less relative URLs, both accepted by
 // java.net.URI but rejected by strict RFC 3986.
 TEST_F(ParseUrlTest, emptyAndRelativeUrls) {
-  const auto parseUrl = [&](const std::optional<std::string>& url,
-                            const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part), url);
-  };
-
   // java.net.URI accepts the empty string; the path is present but empty.
   EXPECT_EQ("", parseUrl("", "PATH"));
   EXPECT_EQ("", parseUrl("", "FILE"));
@@ -184,12 +182,6 @@ TEST_F(ParseUrlTest, emptyAndRelativeUrls) {
 // A second '#' inside the fragment makes the whole URI invalid, so every
 // part is null. A single '#' with an empty fragment is fine.
 TEST_F(ParseUrlTest, doubleHashFragment) {
-  const auto parseUrl = [&](const std::optional<std::string>& url,
-                            const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part), url);
-  };
-
   const auto spaUrl =
       "https://spa.example.com/#/page/orders#/order/unprocessed";
   EXPECT_EQ(std::nullopt, parseUrl(spaUrl, "HOST"));
@@ -217,12 +209,6 @@ TEST_F(ParseUrlTest, doubleHashFragment) {
 // Covers mismatch categories D (non-ASCII characters) and E (brackets).
 // java.net.URI allows unencoded non-ASCII characters and '[]' in queries.
 TEST_F(ParseUrlTest, nonAsciiAndSpecialChars) {
-  const auto parseUrl = [&](const std::optional<std::string>& url,
-                            const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part), url);
-  };
-
   const auto cjkUrl = "https://api.example.com/search?q=\u7f8e\u56e2&userid=1";
   EXPECT_EQ("api.example.com", parseUrl(cjkUrl, "HOST"));
   EXPECT_EQ("/search", parseUrl(cjkUrl, "PATH"));
@@ -274,12 +260,6 @@ TEST_F(ParseUrlTest, nonAsciiAndSpecialChars) {
 // An empty authority yields a null HOST rather than an empty string, and
 // 'http://' with nothing else is invalid.
 TEST_F(ParseUrlTest, emptyAuthority) {
-  const auto parseUrl = [&](const std::optional<std::string>& url,
-                            const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part), url);
-  };
-
   const auto fileUrl = "file:///C:/Users/x/AppData/Temp/%E4%BC%81.png";
   EXPECT_EQ(std::nullopt, parseUrl(fileUrl, "HOST"));
   EXPECT_EQ("file", parseUrl(fileUrl, "PROTOCOL"));
@@ -310,12 +290,6 @@ TEST_F(ParseUrlTest, emptyAuthority) {
 // Pins java.net.URI's authority parsing quirks that surfaced in the
 // mismatch analysis: server-based vs registry-based authority fallback.
 TEST_F(ParseUrlTest, javaAuthorityEdgeCases) {
-  const auto parseUrl = [&](const std::optional<std::string>& url,
-                            const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part), url);
-  };
-
   // Port parsing.
   EXPECT_EQ("h", parseUrl("http://h:", "HOST"));
   EXPECT_EQ("h:", parseUrl("http://h:", "AUTHORITY"));
@@ -403,12 +377,6 @@ TEST_F(ParseUrlTest, javaAuthorityEdgeCases) {
 
 // Pins java.net.URI's scheme and component-presence edge cases.
 TEST_F(ParseUrlTest, javaSchemeAndComponentEdgeCases) {
-  const auto parseUrl = [&](const std::optional<std::string>& url,
-                            const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part), url);
-  };
-
   // The scheme is returned with its original casing and must start with a
   // letter.
   EXPECT_EQ("HTTP", parseUrl("HTTP://x/p", "PROTOCOL"));
@@ -454,11 +422,6 @@ TEST_F(ParseUrlTest, javaSchemeAndComponentEdgeCases) {
 // Pins the three-argument form: the part must be exactly QUERY and the
 // value is extracted with the (&|^)key=([^&]*) regex over the raw query.
 TEST_F(ParseUrlTest, queryKeyExtraction) {
-  const auto parseUrlQuery = [&](const std::optional<std::string>& url,
-                                 const std::optional<std::string>& key) {
-    return evaluateOnce<std::string>("parse_url(c0, 'QUERY', c1)", url, key);
-  };
-
   EXPECT_EQ("1", parseUrlQuery("http://h/p?a=1&&b=2", "a"));
   EXPECT_EQ("2", parseUrlQuery("http://h/p?a=1&&b=2", "b"));
   // An empty key is not special-cased: the regex becomes (&|^)=([^&]*).
@@ -586,12 +549,6 @@ TEST_F(ParseUrlTest, nullArguments) {
 // Pins the '#'-before-'?' split: a '?' inside the fragment belongs to the
 // fragment and never introduces a query.
 TEST_F(ParseUrlTest, questionMarkInFragment) {
-  const auto parseUrl = [&](const std::string& url, const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part),
-        std::optional<std::string>(url));
-  };
-
   const auto url = "http://h/p#f?g";
   EXPECT_EQ(std::nullopt, parseUrl(url, "QUERY"));
   EXPECT_EQ("f?g", parseUrl(url, "REF"));
@@ -608,13 +565,6 @@ TEST_F(ParseUrlTest, questionMarkInFragment) {
 
 // Pins that the key is a regex fragment, not a literal: no escaping.
 TEST_F(ParseUrlTest, regexMetacharacterKey) {
-  const auto parseUrlQuery = [&](const std::string& url,
-                                 const std::string& key) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, 'QUERY', '{}')", key),
-        std::optional<std::string>(url));
-  };
-
   // '.' matches any character, so 'a.b' hits 'aXb=1'.
   EXPECT_EQ("1", parseUrlQuery("http://h/p?aXb=1", "a.b"));
   EXPECT_EQ(std::nullopt, parseUrlQuery("http://h/p?ab=1", "a.b"));
@@ -659,12 +609,6 @@ TEST_F(ParseUrlTest, ipv6LiteralShapes) {
 // Pins that the host keeps its original casing (java.net.URI does not
 // lowercase) and that userinfo escapes pass through undecoded.
 TEST_F(ParseUrlTest, hostCaseAndUserInfoEscapes) {
-  const auto parseUrl = [&](const std::string& url, const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part),
-        std::optional<std::string>(url));
-  };
-
   EXPECT_EQ("HoSt", parseUrl("HTTP://HoSt/p", "HOST"));
   EXPECT_EQ("HoSt", parseUrl("http://HoSt/p", "HOST"));
   EXPECT_EQ("us%65r", parseUrl("http://us%65r@h/", "USERINFO"));
@@ -675,11 +619,6 @@ TEST_F(ParseUrlTest, hostCaseAndUserInfoEscapes) {
 // a stopper at index 40, inside the second chunk (bytes 32..63). All
 // expectations verified against Spark 3.5.1.
 TEST_F(ParseUrlTest, parseUrlLongComponents) {
-  const auto parseUrl = [&](const std::string& url, const std::string& part) {
-    return evaluateOnce<std::string>(
-        fmt::format("parse_url(c0, '{}')", part),
-        std::optional<std::string>(url));
-  };
   const auto repeat = [](int count, char c) { return std::string(count, c); };
   const auto repeatStr = [](std::string_view unit, int count) {
     std::string out;
